+++ src/consts.ts
@@ -1,3 +1,3 @@
-export const CHAIN_SCHEMA_REF = '# yaml-language-server: $schema=../schema.json';
+export const SCHEMA_REF = '# yaml-language-server: $schema=../schema.json';
 export const DEFAULT_GITHUB_REGISTRY = 'https://github.com/hyperlane-xyz/hyperlane-registry';
 export const GITHUB_FETCH_CONCURRENCY_LIMIT = 5;